requirements.txt
# GreenSight - Dependencies

# Google Earth Engine
earthengine-api>=0.1.370
geemap>=0.30.0

# Geospatial
rasterio>=1.3.0
geopandas>=0.14.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Science
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Config & Utils
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.14.0
